refactor: split shell snapshot models
